moved modal to a new file
